refactor (Lotto) : Integer 대신 Number를 참조하도록 리팩토링 한다
