+++ src/main/java/lotto/domain/Lotto.java
@@ -7,28 +7,28 @@ import lotto.util.ValidationUtils;
 
 public class Lotto {
 
-    private final List<Integer> numbers;
+    private final List<Number> numbers;
     private final int LOTTO_SIZE_CRITERION = 6;
 
-    public Lotto(List<Integer> numbers) {
+    public Lotto(List<Number> numbers) {
         validate(numbers);
         this.numbers = numbers;
     }
 
-    private void validate(List<Integer> numbers) {
+    private void validate(List<Number> numbers) {
         ValidationUtils.validateNotNull(numbers);
         validateLottoNumbersSize(numbers);
         validateNoDuplicatedLottoNumbers(numbers);
     }
 
-    private void validateLottoNumbersSize(List<Integer> numbers) {
+    private void validateLottoNumbersSize(List<Number> numbers) {
         if (numbers.size() != LOTTO_SIZE_CRITERION) {
             throw new IllegalArgumentException();
         }
     }
 
-    private void validateNoDuplicatedLottoNumbers(List<Integer> numbers) {
-        Set<Integer> set = new HashSet<>(numbers);
+    private void validateNoDuplicatedLottoNumbers(List<Number> numbers) {
+        Set<Number> set = new HashSet<>(numbers);
         if (set.size() != LOTTO_SIZE_CRITERION) {
             throw new IllegalArgumentException();
         }
