+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,11 +1,11 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
+import static lotto.domain.Numbers.createNumbersList;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @DisplayName("Lotto 클래스")
@@ -14,14 +14,14 @@ class LottoTest {
     @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
     @Test
     void createLottoByOverSize() {
-        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7)))
+        assertThatThrownBy(() -> new Lotto(createNumbersList(List.of(1, 2, 3, 4, 5, 6, 7))))
             .isInstanceOf(IllegalArgumentException.class);
     }
 
     @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
     @Test
     void createLottoByDuplicatedNumber() {
-        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
+        assertThatThrownBy(() -> new Lotto(createNumbersList(List.of(1, 2, 3, 4, 5, 5))))
             .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -30,6 +30,5 @@ class LottoTest {
     void createLottoByNull() {
         assertThatThrownBy(() -> new Lotto(null))
             .isInstanceOf(IllegalArgumentException.class);
-
     }
 }
